% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndividualAAComposition2df.R
\name{calculateIndividualAAComposition2df}
\alias{calculateIndividualAAComposition2df}
\title{calculates proteins aa compositions and returns the results as a data frame}
\usage{
calculateIndividualAAComposition2df(proteins, proteins_subset = NULL)
}
\arguments{
\item{proteins}{a data frame loaded with the readProteinsFasta2df function}

\item{proteins_subset}{an array of protein names to subset from the dataset provided as the proteins parameter (default = NULL)}
}
\value{
a datat_frame composed of the protName as first column and the composition of each AA in the next columns
}
\description{
calculates proteins aa compositions and returns the results as a data frame.
}
