% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProteinsFasta2df.R
\name{readProteinsFasta2df}
\alias{readProteinsFasta2df}
\title{reads a proteins fasta file as a data_frame}
\usage{
readProteinsFasta2df(fasta_file)
}
\arguments{
\item{fasta_file}{a path to the FASTA file}
}
\value{
a datat_frame composed of two columns protName with the preotein names and protSeq containing the fasta sequence
}
\description{
reads a proteins fasta file as a data_frame for further protein wise analyses
}
